+++ CHANGELOG.rst
@@ -23,6 +23,7 @@ Added
 -  Unrolled_recon.py script for training unrolled algorithms.
 - ``train_unrolled.py`` script for training unrolled algorithms.
 - ``benchmark_recon.py`` script for benchmarking and comparing reconstruction algorithms.
+- Added ``reconstruction_error`` to ``ReconstructionAlgorithm`` .
 
 Changed
 ~~~~~~~
@@ -34,8 +35,9 @@ Changed
 -  Updated to Pycsou V2, as ``pip install pycsou`` (Pycsou V1) may not work on some machines.
 -  Data and PSF are now always stored as 4D Data [depth, width, height, color]. Grayscale data has a color axis of length 1 and 2D data has a depth axis of length 1.
 -  Added batch support to RealFFTConvolve2D.
--  Added batch support to RealFFTConvolve2D.
 -  ``ReconstructionAlgorithm.update`` now take the number of the current iteration to allow for unrolled algorithms.
+-  ``ReconstructionAlgorithm.apply`` now takes a reset parameter (default true) to automaticaly call reset.
+-  Call to reset in ``ReconstructionAlgorithm.__init__`` is now optional (see reset parameter).
 
 Bugfix
 ~~~~~~